refactor(vercelclient): extract VercelClient into a new service
